% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSE.R
\name{loadSE}
\alias{loadSE}
\title{Load Example SummarizedExperiment Object  
 
This function loads the example SummarizedExperiment object from the package.  
The example file should be located in the `extdata` directory of the package.}
\value{
A SummarizedExperiment object containing example data.
}
\description{
Load Example SummarizedExperiment Object  
 
This function loads the example SummarizedExperiment object from the package.  
The example file should be located in the `extdata` directory of the package.
}
\examples{
 
library(SEtoolbox)  
example_SE <- loadSE()  
# Now you can analyze the example_SE object  
}
