% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSElist.R
\name{loadSElist}
\alias{loadSElist}
\title{Load Example SummarizedExperiment Object List
 
This function loads the example SummarizedExperiment object list from the package.  
The example file should be located in the `extdata` directory of the package.}
\usage{
loadSElist()
}
\value{
A SummarizedExperiment object list containing example data.
}
\description{
Load Example SummarizedExperiment Object List
 
This function loads the example SummarizedExperiment object list from the package.  
The example file should be located in the `extdata` directory of the package.
}
\examples{
 
library(SEtoolbox)  
SElist <- loadSElist()  
# Now you can analyze the example SE object  
}
