% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_AUCcalc.R
\name{SE_AUCcalc}
\alias{SE_AUCcalc}
\title{Calculate AUC based on gene expressions in a SummarizedExperiment object}
\usage{
SE_AUCcalc(
  SE,
  assayname = "TPM",
  group_colname = "group",
  feature_of_interest,
  maxfeaturecount = 2
)
}
\arguments{
\item{SE}{A SummarizedExperiment object containing gene expression data
and associated metadata.}

\item{assayname}{A string specifying which assay to use from the
SummarizedExperiment object (default is "TPM").}

\item{group_colname}{A string specifying the name of the column
in the metadata that defines the grouping of samples (default is "group").}

\item{feature_of_interest}{A data.frame containing at least two columns:
`geneList` (the gene names) and `geneWeight` (the weights for each gene).}

\item{maxfeaturecount}{An integer specifying the maximum number of features
(genes) to consider in the combinations (default is 2).}
}
\value{
A data.frame containing the best AUC results for the combinations   
  of genes tested. Each row corresponds to a particular combination of genes   
  along with its AUC score.
}
\description{
This function calculates the Area Under the Curve (AUC) for various combinations   
of genes in the provided SummarizedExperiment (SE) object. It allows for weighted   
genes and enables users to specify the maximum number of features to combine.
}
