% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_boxplot.R
\name{SE_boxplot}
\alias{SE_boxplot}
\title{SE_boxplot: Create Boxplots for SummarizedExperiment Object}
\usage{
SE_boxplot(
  SE,
  feature_of_interest = c("AAGAB", "ABCA13", "ABCC4", "ABHD2"),
  assayname = "TPM",
  group_colname = "group",
  normalization = "none"
)
}
\arguments{
\item{SE}{A \code{SummarizedExperiment} object that contains gene expression data.}

\item{feature_of_interest}{A character vector specifying the gene names to be plotted. 
Defaults to \code{c("AAGAB", "ABCA13", "ABCC4", "ABHD2")}.}

\item{assayname}{A string indicating the assay from which to extract the data. 
The default value is \code{"TPM"}.}

\item{group_colname}{A string representing the column name in \code{colData} that 
holds group information. Defaults to \code{"group"}.}

\item{normalization}{A string specifying the normalization method. 
Available options are \code{"none"}, \code{"scale"}, or \code{"log"}. Defaults to \code{"none"}.}
}
\value{
A \code{ggplot} object that represents the violin and box plots, including significance markers.
}
\description{
This function generates violin and box plots for specified genes within a 
\code{SummarizedExperiment} object. It enables the visualization of gene 
expression levels across different groups. Optionally, it can perform data 
normalization and uses ANOVA with Tukey's post-hoc test to add significance 
markers on the generated plot.
}
\examples{
# Create a dummy SummarizedExperiment object
data_matrix <- matrix(rnorm(1000), nrow = 100, ncol = 10)
rownames(data_matrix) <- paste0("Gene", 1:100)
colnames(data_matrix) <- paste0("Sample", 1:10)
sample_info <- DataFrame(group = rep(c("A", "B"), each = 5))
SE <- SummarizedExperiment(assays = list(TPM = data_matrix), colData = sample_info)

# Call the SE_boxplot function
plot <- SE_boxplot(SE, feature_of_interest = c("Gene1", "Gene2"), 
                   group_colname = "group", normalization = "log")
print(plot)

}
