% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_combat.R
\name{SE_combat}
\alias{SE_combat}
\title{Batch Effect Correction for SummarizedExperiment Objects  
 
This function applies the ComBat algorithm to correct batch effects in   
a SummarizedExperiment object based on a specified batch variable.}
\usage{
SE_combat(SE, col_for_combat)
}
\arguments{
\item{SE}{A SummarizedExperiment object containing the assay data.}

\item{col_for_combat}{A character string specifying the column name of
the batch variable in the sample data. This must be provided and valid.}
}
\value{
A SummarizedExperiment object with the assay data corrected for   
  batch effects.
}
\description{
Batch Effect Correction for SummarizedExperiment Objects  
 
This function applies the ComBat algorithm to correct batch effects in   
a SummarizedExperiment object based on a specified batch variable.
}
