% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_combine.R
\name{SE_combine}
\alias{SE_combine}
\title{Combine Multiple SummarizedExperiment Objects  
 
This function merges multiple SummarizedExperiment objects based on the specified merge type (intersection or union) for all assays present in the input list.}
\usage{
SE_combine(se_list, merge_type = "intersection")
}
\arguments{
\item{se_list}{A list of SummarizedExperiment objects to be combined.}

\item{merge_type}{A character string specifying the type of merge to perform. Options are "intersection" (default) to keep only common features across all objects, or "union" to keep all features.}
}
\value{
A SummarizedExperiment object containing the combined data for all assays.
}
\description{
Combine Multiple SummarizedExperiment Objects  
 
This function merges multiple SummarizedExperiment objects based on the specified merge type (intersection or union) for all assays present in the input list.
}
\examples{
 
# Example usage  
se1 <- SummarizedExperiment(assays = list(counts = matrix(rnorm(100), nrow = 10)), colData = DataFrame(sample = 1:10))  
se2 <- SummarizedExperiment(assays = list(counts = matrix(rnorm(100), nrow = 10)), colData = DataFrame(sample = 11:20))  
combined_se <- SE_combine(list(se1, se2), merge_type = "union")  
 
}
