% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_detectratio.R
\name{SE_detectratio}
\alias{SE_detectratio}
\title{SE_detectratio: Calculate Detection Ratio for SummarizedExperiment Object}
\usage{
SE_detectratio(SE, assayname = "TPM", group_colname = NULL)
}
\arguments{
\item{SE}{A \code{SummarizedExperiment} object containing gene expression data. This
should include the assay for which detection ratios will be calculated.}

\item{assayname}{A string indicating which assay to use for calculations.
The default value is \code{"TPM"}. The assay must exist in the \code{SummarizedExperiment} object.}

\item{group_colname}{A string representing the column name in \code{colData} that
contains group information. The default value is \code{"group"}. This is optional; if not provided,   
the function will calculate overall detection ratios only.}
}
\value{
A list containing:  
\item{SE}{An updated \code{SummarizedExperiment} object with detection sample counts  
and ratios added to \code{rowData}. The detection sample counts are added to the   
\code{"detectsample"} and corresponding ratios to the \code{"detectratio"} columns.}  
\item{plot_feature}{A \code{ggplot} object representing the histogram of detection ratios   
for features. This visualizes the distribution of detection ratios across all features   
and groups if applicable.}  
\item{plot_sample}{A \code{ggplot} object representing the bar plot of detection ratios   
for samples. The Y-axis displays the expression fraction formatted as a percentage,   
and the samples are displayed on the X-axis.}
}
\description{
This function computes the detection ratio of expression data in a   
\code{SummarizedExperiment} object and updates its \code{rowData} with detection   
sample counts and ratios. It also generates a histogram of detection ratios to   
visualize the distribution. Detection ratios are calculated based on the number   
of non-zero samples for each feature across provided groups. The function can also   
calculate detection ratios per group if specified.
}
\examples{
  
# Load example SummarizedExperiment object 

SE <- loadSE()   

# Call the SE_detectratio function
 
result <- SE_detectratio(SE, assayname = "TPM", group_colname = "group")   

SEdectectratio <-  result$SE # Return an SummarizedExperiment object after detectratio calculate 
 
print(result$plot_sample)  # Display the sample histogram  

print(result$plot_feature)  # Display the sample histogram
 
}
