% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_distribution.R
\name{SE_distribution}
\alias{SE_distribution}
\title{Plot Distribution of Non-Zero Entries in an Expression Set  
 
This function generates two plots: a bar plot showing the count of non-zero entries   
for each feature in the given expression matrix, and a histogram representing the   
distribution of the fraction of non-zero entries across samples.}
\usage{
SE_distribution(SE, assayname = "TPM", ZeroasNA = TRUE)
}
\arguments{
\item{SE}{A SummarizedExperiment object containing the expression data.}

\item{assayname}{A string indicating the assay name to use from the SummarizedExperiment object. Default is "TPM".}

\item{ZeroasNA}{A logical value indicating whether zeros should be treated as NA. Default is TRUE.}
}
\value{
A ggplot object containing a combined plot of the bar plot and histogram.
}
\description{
Plot Distribution of Non-Zero Entries in an Expression Set  
 
This function generates two plots: a bar plot showing the count of non-zero entries   
for each feature in the given expression matrix, and a histogram representing the   
distribution of the fraction of non-zero entries across samples.
}
\examples{
 
# Assuming `se` is a SummarizedExperiment object  
distribution_plot <- SE_distribution(SE = se, assayname = "TPM", ZeroasNA = TRUE)  
print(distribution_plot)  
 
}
