% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_heatmap.R
\name{SE_heatmap}
\alias{SE_heatmap}
\title{Calculate the Detection Proportion of Expression Data  
 
This function computes the detection proportion of expression data   
contained within a SummarizedExperiment object.}
\usage{
SE_heatmap(
  SE,
  assayname = "TPM",
  group_col = NULL,
  normalization = "none",
  genes_of_interest = NULL,
  use_raster = TRUE
)
}
\arguments{
\item{SE}{A SummarizedExperiment object containing the expression data
and sample information.}

\item{assayname}{A character string specifying the assay name of the
expression matrix. Default is "TPM".}

\item{group_col}{An optional character string specifying the column name
in the sample metadata to group the samples. If NULL, no grouping will   
be applied.}

\item{normalization}{A character string specifying the normalization method
to be used. Default is "none". Options may include "log" or "scale".}

\item{genes_of_interest}{An optional vector of gene names to focus on
in the analysis. If NULL, all genes will be included.}

\item{use_raster}{A logical value indicating whether to use rasterization
for heatmap plotting. Default is TRUE.}
}
\value{
A SummarizedExperiment object containing updated information on  
  the detection proportions.
}
\description{
Calculate the Detection Proportion of Expression Data  
 
This function computes the detection proportion of expression data   
contained within a SummarizedExperiment object.
}
