% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_PCAplot.R
\name{SE_PCAplot}
\alias{SE_PCAplot}
\title{生成PCA图}
\usage{
SE_PCAplot(
  SE,
  assayname = "TPM",
  groupname = "group",
  outlier_threshold = 2,
  scale = TRUE
)
}
\arguments{
\item{SE}{SummarizedExperiment对象，包含基因表达数据。}

\item{assayname}{指定的表达数据名称，默认为"TPM"。}

\item{groupname}{分组列名称，默认为"group"。}

\item{outlier_threshold}{离群值过滤阈值，默认为2。}

\item{scale}{是否对数据进行标准化，默认为TRUE。}
}
\value{
包含两个ggplot对象的列表，原始PCA图和过滤后的PCA图。
}
\description{
该函数接受一个SummarizedExperiment对象，计算PCA并对结果进行可视化。
}
