% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_boxplot.R
\name{SE_boxplot}
\alias{SE_boxplot}
\title{Generates a boxplot for specified features in a SummarizedExperiment object.}
\usage{
SE_boxplot(
  SE,
  feature_of_interest = c("AAGAB", "ABCA13", "ABCC4", "ABHD2"),
  assayname = "TPM",
  group_col = NA,
  normalization = "none"
)
}
\arguments{
\item{SE}{A SummarizedExperiment object.}

\item{feature_of_interest}{Character vector of gene identifiers.}

\item{assayname}{The assay name in the SummarizedExperiment object.}

\item{group_col}{Column name in colData(SE) for grouping.}

\item{normalization}{Normalization method ("scale", "log", or "none").}
}
\description{
Generates a boxplot for specified features in a SummarizedExperiment object.
}
