% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_circTest.R
\name{SE_circTest}
\alias{SE_circTest}
\title{Circular RNA Differential Expression Test Using Beta - Binomial Model}
\usage{
SE_circTest(SECirc, SELinear, assayname = "Count", group_colname = "group")
}
\arguments{
\item{SECirc}{A SummarizedExperiment object that holds circular RNA count data.}

\item{SELinear}{A SummarizedExperiment object that holds linear RNA count data.}

\item{assayname}{Character string specifying the name of the assay containing count data. Defaults to "Count".}

\item{group_colname}{Character string indicating the name of the column in \code{colData(SECirc)}
that provides group - related information. Defaults to "group".}
}
\value{
A data frame with circRNA names as row names. It contains columns for:
        - Mean expression ratios in each group.
        - Raw p - values obtained from likelihood ratio tests.
        - Adjusted p - values using the Benjamini - Hochberg method.
}
\description{
This function conducts differential expression analysis for circular RNAs (circRNAs).
It compares the expression ratios of circRNAs among different groups by utilizing a beta - binomial model.
}
\details{
The function first extracts common features and samples between circular and linear RNA data.
Then, it calculates the ratio of circular RNA expression to the total (circular + linear) RNA expression
for each circRNA. After that, it fits null and alternative beta - binomial models to test for significant
differences in these ratios between groups, addressing overdispersion in count data.
}
\examples{
\dontrun{
# Example usage
results <- SE_circTest(SECirc = myCircSE, SELinear = myLinearSE, 
                       group_colname = "treatment_group")
}

}
