% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_circTest.R
\name{SE_circTest}
\alias{SE_circTest}
\title{Circular RNA Differential Expression Test Using Beta - Binomial Model}
\usage{
SE_circTest(
  SEcirc,
  SElinear,
  assayname = "Count",
  group_colname = "group",
  select_group = NULL
)
}
\arguments{
\item{SEcirc}{A SummarizedExperiment object that holds circular RNA count data.}

\item{SElinear}{A SummarizedExperiment object that holds linear RNA count data.}

\item{assayname}{Character string specifying the name of the assay containing count data. Defaults to "Count".}

\item{group_colname}{Character string indicating the name of the column in \code{colData(SECirc)}
that provides group - related information. Defaults to "group".}

\item{select_group}{Character vector of groups to include, eg c("control","case"); NULL includes all groups (default: NULL).}
}
\value{
A data frame with circRNA names as row names. It contains columns for:
        - Mean expression ratios in each group.
        - Raw p - values obtained from likelihood ratio tests.
        - Adjusted p - values using the Benjamini - Hochberg method.
        - The data frame is sorted by raw p - values in ascending order.
}
\description{
This function conducts differential expression analysis for circular RNAs (circRNAs).
It compares the expression ratios of circRNAs among different groups by utilizing a beta - binomial model.
The function also addresses overdispersion in count data and provides adjusted p - values for multiple testing.
}
\details{
The function first extracts common features and samples between circular and linear RNA data.
Then, it calculates the ratio of circular RNA expression to the total (circular + linear) RNA expression
for each circRNA. After that, it fits null and alternative beta - binomial models to test for significant
differences in these ratios between groups. Finally, it adjusts the p - values for multiple testing
using the Benjamini - Hochberg method and sorts the results by raw p - values.
}
\examples{
\dontrun{
# Example usage
results <- SE_circTest(SEcirc = SEcirc, SElinear = SElinear, 
                       group_colname = "group", select_group = c("control", "case"))
}

}
