% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_COCONUT.R
\name{SE_COCONUT}
\alias{SE_COCONUT}
\title{Apply COCONUT method for batch effect removal on SummarizedExperiment objects}
\usage{
SE_COCONUT(
  SElist,
  assayname = "TPM",
  group_col = "group",
  label_healthy = "HC"
)
}
\arguments{
\item{SElist}{A list of SummarizedExperiment objects that contain expression data and sample information.}

\item{assayname}{The name of the assay in the SummarizedExperiment objects to be used for analysis. Default is "TPM".}

\item{group_col}{The column name in the sample information that indicates the group of each sample. Default is "group".}

\item{label_healthy}{The label in the `group_col` that represents the healthy samples. Default is "HC".}
}
\value{
A new SummarizedExperiment object with batch - corrected expression data.
}
\description{
This function takes a list of SummarizedExperiment objects and applies the COCONUT method
to remove batch effects. It creates COCO objects from the input SummarizedExperiment objects,
processes them using COCONUT, and then combines the processed expression data to create a
new SummarizedExperiment object with batch - corrected data.
}
\examples{
# Assume selist is a list of SummarizedExperiment objects
# SE_corrected <- SE_COCONUT(selist)

}
