% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_combat.R
\name{SE_combat}
\alias{SE_combat}
\title{Perform Batch Effect Correction using ComBat  
 
This function adjusts for batch effects in a SummarizedExperiment object using the ComBat method.}
\usage{
SE_combat(SE, col_for_combat, col_for_compare = NULL)
}
\arguments{
\item{SE}{A SummarizedExperiment object containing expression data.}

\item{col_for_combat}{A string specifying the column name in the sample information that contains batch information.}

\item{col_for_compare}{An optional string specifying the column name in the sample information that contains the condition for comparison.}
}
\value{
A SummarizedExperiment object with batch effects corrected.
}
\description{
Perform Batch Effect Correction using ComBat  
 
This function adjusts for batch effects in a SummarizedExperiment object using the ComBat method.
}
\examples{
 
# Assuming 'se' is a SummarizedExperiment object  
corrected_se <- SE_combat(se, col_for_combat = "batch", col_for_compare = "condition")  
 
}
