% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_DESeq2.R
\name{SE_DEseq2}
\alias{SE_DEseq2}
\title{Perform differential expression analysis using DESeq2}
\usage{
SE_DEseq2(SE, assayname = "Count", groupname = "group")
}
\arguments{
\item{SE}{A SummarizedExperiment object containing count data.}

\item{assayname}{The name of the assay to use for the analysis. Default is "Count".}

\item{groupname}{The name of the column in colData(SE) that contains the factor for grouping samples. Default is "group".}
}
\value{
A SummarizedExperiment object with DESeq2 results stored in its metadata.
}
\description{
Perform differential expression analysis using DESeq2
}
