% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_DESeq2.R
\name{SE_DEseq2}
\alias{SE_DEseq2}
\title{Perform differential expression analysis using DESeq2}
\usage{
SE_DEseq2(SE, assayname = "Count", group_colname = "group")
}
\arguments{
\item{SE}{A SummarizedExperiment object containing count data.}

\item{assayname}{The name of the assay to use for the analysis. Default is "Count".}

\item{group_colname}{The name of the column in colData(SE) that contains the factor for grouping samples. Default is "group".}
}
\value{
A SummarizedExperiment object with DESeq2 results stored in its metadata.
}
\description{
This function conducts differential expression analysis on a given SummarizedExperiment object  
using the DESeq2 package. It takes count data from the specified assay and performs  
two-group comparisons based on the grouping factor provided in the colData.
}
\examples{
 
# Load necessary libraries  
library(SummarizedExperiment)  
library(DESeq2)  
 
# Load example SummarizedExperiment object  
SE = loadSE() 
 
# Perform differential expression analysis  
results_se <- SE_DEseq2(SE = SE, assayname = "Count", group_colname = "group")  
 
}
