% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_detectratio.R
\name{SE_detectratio}
\alias{SE_detectratio}
\title{计算检测比例并更新SummarizedExperiment对象的rowData}
\usage{
SE_detectratio(SE, assayname = "TPM")
}
\arguments{
\item{SE}{SummarizedExperiment对象}

\item{assayname}{用于计算的assay名称，默认为"TPM"}
}
\value{
更新了detectsample和detectratio列的SummarizedExperiment对象
}
\description{
计算检测比例并更新SummarizedExperiment对象的rowData
}
