% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_heatmap.R
\name{SE_heatmap}
\alias{SE_heatmap}
\title{计算表达数据的检测比例}
\usage{
SE_heatmap(
  SE,
  assayname = "TPM",
  group_col = NULL,
  normalization = "none",
  genes_of_interest = NULL,
  use_raster = TRUE
)
}
\arguments{
\item{SE}{SummarizedExperiment对象，包含表达数据和样本信息。}

\item{assayname}{表达矩阵的assay名称，默认为"TPM"}
}
\value{
返回包含检测比例信息的SummarizedExperiment对象
}
\description{
计算表达数据的检测比例
}
