% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_impute.R
\name{SE_impute}
\alias{SE_impute}
\title{Impute missing values in a SummarizedExperiment object}
\usage{
SE_impute(SE, assayname = "TPM", group = "group", method = "knn")
}
\arguments{
\item{SE}{A SummarizedExperiment object containing expression data.}

\item{assayname}{The name of the expression data matrix to be imputed. Default is "TPM".}

\item{group}{The column name of sample grouping information. Currently not used in this function. Default is "group".}

\item{method}{The imputation method. Currently only "knn" is supported. Default is "knn".}
}
\value{
A new SummarizedExperiment object with imputed expression data.
}
\description{
This function is used to impute missing values in the expression data of a SummarizedExperiment object.
It replaces zero values in the expression data with NA and then performs imputation using the kNN method.
}
\examples{
# Assume se is a SummarizedExperiment object
# SEimpute <- SE_impute(se)

}
