% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_impute.R
\name{SE_impute}
\alias{SE_impute}
\title{填补 SummarizedExperiment 对象中的缺失值  
 
本函数利用指定的方法对给定的 SummarizedExperiment 对象中的缺失值 (NA) 进行填补。  
可以使用多种插补技术来处理缺失值，以确保后续分析的稳健性。}
\usage{
SE_impute(
  object,
  assayname = "TPM",
  group = "group",
  ZerosAsNA = FALSE,
  RemoveNA = TRUE,
  cutoff = 20,
  method = c("none", "LOD", "half_min", "median", "mean", "min", "knn", "rf",
    "global_mean", "svd", "QRILC"),
  LOD = NULL,
  knum = 10
)
}
\arguments{
\item{object}{一个 SummarizedExperiment 对象，包含需插补的数据。}

\item{assayname}{SummarizedExperiment assay名字，选择待插补数据类型。}

\item{group}{一个字符字符串，指定样本数据中的分组变量。}

\item{ZerosAsNA}{一个逻辑值，指示是否将零视为 NA。默认值为 FALSE。}

\item{RemoveNA}{一个逻辑值，指示是否基于 cutoff 删除高 NA 百分比的样本。默认值为 TRUE。}

\item{cutoff}{一个数值，表示 NA 样本的百分比截断值。默认值为 20。}

\item{method}{一个字符字符串，指定使用的插补方法。可选项包括 "none"、"LOD"、"half_min"、
"median"、"mean"、"min"、"knn"、"rf"、"global_mean"、"svd" 和 "QRILC"。默认值为 "none"。}

\item{LOD}{一个数值，表示检出限（用于 LOD 插补方法）。默认值为 NULL。}

\item{knum}{一个整数值，表示 KNN 插补方法中邻居的数量。默认值为 10。}
}
\value{
SummarizedExperiment 对象，包含填补后的值。
}
\description{
填补 SummarizedExperiment 对象中的缺失值  
 
本函数利用指定的方法对给定的 SummarizedExperiment 对象中的缺失值 (NA) 进行填补。  
可以使用多种插补技术来处理缺失值，以确保后续分析的稳健性。
}
\examples{
 
# 使用示例  
# 假设 `se` 是一个 SummarizedExperiment 对象  
se_imputed <- SE_impute(se, assayname = "TPM", group = "my_group_column", method = "median", ZerosAsNA = TRUE)  
 
}
