% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_PCAplot.R
\name{SE_PCAplot}
\alias{SE_PCAplot}
\title{Generate PCA plots  
 
This function takes a SummarizedExperiment object, computes PCA, and visualizes the results.}
\usage{
SE_PCAplot(
  SE,
  assayname = "TPM",
  groupname = "group",
  outlier_threshold = 2,
  scale = TRUE,
  feature_of_interesting = NULL,
  show_caption = TRUE
)
}
\arguments{
\item{SE}{SummarizedExperiment object containing gene expression data.}

\item{assayname}{Name of the expression data, default is "TPM".}

\item{groupname}{Name of the grouping column, default is "group".}

\item{outlier_threshold}{Outlier filtering threshold, default is 2.}

\item{scale}{Whether to standardize the data, default is TRUE.}

\item{feature_of_interesting}{Vector of specific feature names; if NULL, all features are used, default is NULL.}

\item{show_caption}{Logical value indicating whether to display caption, default is TRUE.}
}
\value{
A list containing two ggplot objects: the original PCA plot and the filtered PCA plot.
}
\description{
Generate PCA plots  
 
This function takes a SummarizedExperiment object, computes PCA, and visualizes the results.
}
